Convert back from extension method to regular method for Java interop
