+++ fuel-android/src/main/kotlin/com/github/kittinunf/fuel/android/extension/Requests.kt
@@ -4,10 +4,6 @@ import com.github.kittinunf.fuel.core.*
 import com.github.kittinunf.result.Result
 import org.json.JSONObject
 
-/**
- * Created by Kittinun Vantasin on 11/9/15.
- */
-
 //jsonObject
 public fun Request.responseJson(handler: (Request, Response, Result<JSONObject, FuelError>) -> Unit) =
         response(jsonDeserializer(), handler)
