+++ fuel-android/src/main/kotlin/com/github/kittinunf/fuel/android/util/AndroidEnvironment.kt
@@ -5,20 +5,9 @@ import android.os.Looper
 import com.github.kittinunf.fuel.core.Environment
 import java.util.concurrent.Executor
 
-/**
- * Created by Kittinun Vantasin on 11/9/15.
- */
-
 internal class AndroidEnvironment : Environment {
-
     val handler = Handler(Looper.getMainLooper())
 
-    override var callbackExecutor: Executor = object : Executor {
-
-        override fun execute(command: Runnable?) {
-            handler.post(command)
-        }
-
-    }
+    override var callbackExecutor: Executor = Executor { command -> handler.post(command) }
 
 }
