+++ fuel-android/src/test/kotlin/com/github/kittinunf/fuel/android/BaseTestCase.kt
@@ -6,10 +6,6 @@ import org.robolectric.annotation.Config
 import java.util.concurrent.CountDownLatch
 import java.util.concurrent.TimeUnit
 
-/**
- * Created by Kittinun Vantasin on 11/9/15.
- */
-
 @RunWith(RobolectricGradleTestRunner::class)
 @Config(constants = BuildConfig::class, sdk = intArrayOf(21))
 abstract class BaseTestCase {
